# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user-dashboard/statistics": {
            "get": {
                "tags": [
                    "User Dashboard"
                ],
                "summary": "Get borrow statistics for authenticated user",
                "operationId": "abbab5a50f7bccd8629c6e770a001923",
                "responses": {
                    "200": {
                        "description": "User borrow statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statistics": {
                                            "properties": {
                                                "total_borrowed_books": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total_returned_books": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "total_overdue_books": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-dashboard/borrowed-books": {
            "get": {
                "tags": [
                    "User Dashboard"
                ],
                "summary": "List borrowed books for authenticated user",
                "operationId": "1f86ee279f0a828bc97ff97f7df64209",
                "responses": {
                    "200": {
                        "description": "Paginated list of borrowed books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Borrow"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book Resource",
                "required": [
                    "id",
                    "category",
                    "name",
                    "author",
                    "total_copies",
                    "available_copies"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "description": "Either category ID or full category object if loaded",
                        "oneOf": [
                            {
                                "type": "integer",
                                "example": 2
                            },
                            {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Fiction"
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    },
                    "name": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "author": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                    },
                    "short_description": {
                        "type": "string",
                        "example": "A classic novel set in the 1920s",
                        "nullable": true
                    },
                    "total_copies": {
                        "type": "integer",
                        "example": 10
                    },
                    "available_copies": {
                        "type": "integer",
                        "example": 7
                    },
                    "average_rating": {
                        "type": "integer",
                        "example": 4
                    },
                    "total_ratings": {
                        "type": "integer",
                        "example": 100
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 25
                    },
                    "book_cover_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/book_covers/cover.jpg",
                        "nullable": true
                    },
                    "pdf_file_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/pdf_files/book.pdf",
                        "nullable": true
                    },
                    "audio_file_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://localhost/storage/audio_files/book.mp3",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Borrow": {
                "required": [
                    "id",
                    "borrowed_at",
                    "return_date",
                    "status",
                    "extension_count"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "borrowed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-01T10:00:00Z"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T23:59:59Z"
                    },
                    "status": {
                        "type": "string",
                        "example": "borrowed"
                    },
                    "returned_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-25T14:30:00Z",
                        "nullable": true
                    },
                    "extension_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "book": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Book"
                            }
                        ],
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "customer"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Dhaka, Bangladesh"
                    },
                    "profile_picture": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/users/john.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}